delete most stuff to get the basics nearly working
